move edit component and controls into edit folder
