--- packages/block-library/src/terms-query/edit.js
@@ -1,10 +1,0 @@
-/**
- * Internal dependencies
- */
-import TermsQueryContent from './terms-query-content';
-
-const TermsQueryEdit = ( props ) => {
-	return <TermsQueryContent { ...props } />;
-};
-
-export default TermsQueryEdit;
